WARNING: THIS IS NOT BACKWARDS COMPATIBLE WITH PREVIOUS VERSIONS
To accommodate some features in Rapid 2.2, I had to rewrite the apiWrapper function, create header parsing, a staging of queries and a different retrieval method for the Rapid responses.

class->body() now returns the body of the response
class->header() returns the header.

An additional element $link has been created, populated in case there is a paging result.